requirements.txt
entrypoints==0.4
Flask==2.2.3
GitPython==3.1.31
gunicorn==20.1.0
h5py==3.8.0
keras==2.12.0
ml-dtypes==0.0.4
numpy==1.23.5
pandas==1.5.3
Pillow==9.5.0
requests==2.28.2
requests-oauthlib==1.3.1
streamlit==1.21.0
urllib3==1.26.15
tensorflow-cpu==2.12.0
